+++ src/components/tabs/TabList.svelte
@@ -1,4 +1,4 @@
-<div class="mb-4 flex flex-row">
+<div class=" flex flex-row">
   <slot />
-  <div class="flex-grow border-b border-gray-600" />
+  <div class="flex-grow border-b border-gray-200" />
 </div>
