+++ src/routes/rain-chess/liChessPlayer.svelte
@@ -5,12 +5,9 @@
   import { validateFields } from "../../utils";
   import { addressValidate, required } from "../../validation";
   import Parser from "$components/parser/Parser.svelte";
-  import Label from "$routes/rain-chess/Label.svelte";
-  import Info from "$routes/rain-chess/Info.svelte";
   import Section from "$routes/rain-chess/Section.svelte";
   import SectionHeading from "$routes/rain-chess/SectionHeading.svelte";
   import SectionBody from "$routes/rain-chess/SectionBody.svelte";
-  import Item from "$routes/rain-chess/Item.svelte";
   import { writable, type Writable } from "svelte/store";
   import IconLibrary from "$components/IconLibrary.svelte";
   import StakeNBuy from "$routes/rain-chess/Stake&Buy.svelte";
@@ -23,6 +20,7 @@
 
   import ContractsConfigs from "../../../mumbai.json"
     import { Logger, parseUnits } from "ethers/lib/utils";
+    import UserDetails from "./UserDetails.svelte";
 
   $: oAuth = JSON.parse(localStorage.getItem('oauth2authcodepkce-state'))
   const { open } = getContext('simple-modal') 
@@ -56,19 +54,18 @@
 
   $: if($signer){   
     const authorizeAccount = async () => { 
-
-       try { 
+      try { 
 
         let authToken = JSON.parse(localStorage.getItem('oauth2authcodepkce-state')) 
-       console.log('authToken : ' , authToken)
-       console.log('authToken : ' , authToken.accessToken.value) 
-       oAuth = authToken
+        console.log('authToken : ' , authToken)
+        console.log('authToken : ' , authToken.accessToken.value) 
+        oAuth = authToken
 
-       let authResult = await axios.post('https://gildlab-ipfs.in.ngrok.io/lichess/api/v2/verifyAccount' , {address : $signerAddress , lichessToken : authToken?.accessToken?.value})  
-       isWalletCorrect = true
-       alert(`${authResult.data.message}`)
+        let authResult = await axios.post('https://gildlab-ipfs.in.ngrok.io/lichess/api/v2/verifyAccount' , {address : $signerAddress , lichessToken : authToken?.accessToken?.value})  
+        isWalletCorrect = true
+        alert(`${authResult.data.message}`)
         
-       } catch (error) {
+      } catch (error) {
           console.log("error : " ,error)  
           isWalletCorrect = false
           if(!error?.response?.data?.status && error?.response?.data?.code == 401){
@@ -84,21 +81,7 @@
 
   } 
   
-    const getAccountData = async () =>{
-
-      let tokenData = JSON.parse(localStorage.getItem('oauth2authcodepkce-state')) 
-      
-      let data = await axios.get('https://lichess.org/api/account',{
-        headers: { 
-          'Authorization': `Bearer ${tokenData?.accessToken?.value}`
-        }
-      })
-      console.log("data", data);
-      
-      accDetails = data.data
-    }
     authorizeAccount()
-    getAccountData()
   }
   // setTimeout(getAccountData, 3000);
 
@@ -264,40 +247,6 @@
 
 <div class="flex gap-x-3 relative">
   <div class="flex w-2/3 flex-col gap-y-6 p-8">
-              <!-- {#if isClaimRes} -->
-              <!-- <Section>
-                <SectionHeading>Approve Amount</SectionHeading>
-                <SectionBody>
-                  <Select
-                      items={Options}
-                      bind:value={option}
-                      on:change={() =>{
-                        if(option.value == 1 ){
-                          document.getElementById("approveAmount").style.display = 'grid'
-                        }
-                      }}
-                    >
-                      <span slot="label"> Select The Option: </span>
-                    </Select>
-                  <div id="approveAmount" class="w-full" style="display: none;">
-                    <div class="grid grid-cols-2 gap-4">
-                    <Input
-                      type="text"
-                      placeholder="Enter Amount"
-                      bind:this={fields.approveAmount}
-                      bind:value={approveAmount}
-                      validator={required}
-                    >
-                      <span slot="label">Enter the units to approve</span>
-                    </Input>
-                  </div>
-                  <span class="pt-4">
-                    <Button shrink disabled={!approveAmount} on:click={approve}> Approve </Button>
-                  </span>
-                </div>
-                </SectionBody>
-              </Section> -->
-            <!-- {/if} -->
     {#if !walletVerified }
     <span class="text-3xl font-semibold">liChess Player</span>
         <Section>
@@ -314,93 +263,7 @@
           </SectionBody>
         </Section>
     {:else}  
-
-      {#if $signer}
-        <Section>
-            <SectionHeading>User Details</SectionHeading>
-            <div class="p-3 pl-5">
-              {#if !accDetails?.profile}
-                Username : <span class="uppercase font-semibold"><a target='_blank' class="text-blue-400 underline" href={accDetails?.url}>{accDetails ? accDetails?.username : "John Doe"}</a></span>
-              {:else}
-                Name : <span class="uppercase font-semibold"><a target='_blank' class="text-blue-400 underline" href={accDetails?.url}>{`${accDetails ? (accDetails?.profile?.firstName + " " + accDetails?.profile?.lastName) : "John Doe"}`}</a></span>
-              {/if}
-            </div>
-            <SectionBody>
-              <div class="flex flex-row">
-                <div class="w-1/2">
-              <Item gap="gap-y-4">
-                <Label>Games Details: </Label>
-                <Info>
-                  <span class="flex gap-x-4 ">
-                    <img src="/assets/allGame.png" width='30' height='30' alt='all' class='me-4' /> Total games : {`${accDetails ? accDetails?.count.all : 0}`}
-                  </span>
-                </Info>
-                <Info>
-                  <span class="flex gap-x-4 ">
-                    <img src="/assets/win.png" width='30' height='30' alt='all' class='me-4' /> Win games : {`${accDetails ? accDetails?.count.win : 0}`}
-                  </span>
-                </Info>
-                <Info>
-                  <span class="flex gap-x-4 ">
-                    <img src="/assets/drawn.png" width='30' height='30' alt='all' class='me-4' /> Draw games : {`${accDetails ? accDetails?.count.draw : 0}`}
-                  </span>
-                </Info>
-                <Info>
-                  <span class="flex gap-x-4 ">
-                    <img src="/assets/loss.png" width='30' height='30' alt='all' class='me-4' /> Loss games : {`${accDetails ? accDetails?.count.loss : 0}`}
-                  </span>
-                  </Info>
-                <Info>
-                  <span class="flex gap-x-4 ">
-                    <img src="/assets/playing.png" width='30' height='30' alt='all' class='me-4' /> Playing games : {`${accDetails ? accDetails?.count.playing : 0}`}
-                  </span>
-                </Info>
-              </Item>
-            </div>
-            <div class="w-1/2">
-              <Item>
-                <Label>NFTs : </Label>
-                <div class="flex py-4">
-                  <div class="flex w-1/2 gap-x-4">
-                    <img src="/assets/bullet.png" width='35' height='35' alt='all' class='me-4' />
-                    <div class="flex flex-col">
-                      <div>BULLET</div>
-                      <div>{`${accDetails ? accDetails?.perfs?.bullet?.games : 0} games`}</div>
-                    </div>
-                  </div>
-
-                  <div class="flex w-1/2 gap-x-4">
-                    <img src="/assets/blitz.png" width='40' height='26' alt='all' class='me-4' />
-                    <div class="flex flex-col">
-                      <div>BLITZ</div>
-                      <div>{`${accDetails ? accDetails?.perfs?.blitz?.games : 0} games`}</div>
-                    </div>
-                  </div>
-                </div>
-
-                <div class="flex py-4">
-                  <div class="flex w-1/2 gap-x-4">
-                    <img src="/assets/rapid.png" width='35' height='35' alt='all' class='me-4' />
-                    <div class="flex flex-col">
-                      <div>RAPID</div>
-                      <div>{`${accDetails ? accDetails?.perfs?.rapid?.games : 0} games`}</div>
-                    </div>
-                  </div>
-
-                  <div class="flex w-1/2 gap-x-4">
-                    <img src="/assets/classical.png" width='35' height='35' alt='all' class='me-4' />
-                    <div class="flex flex-col">
-                      <div>CLASSICAL</div>
-                      <div>{`${accDetails ? accDetails?.perfs?.classical?.games : 0} games`}</div>
-                    </div>
-                  </div>
-                </div>
-              </Item>
-            </div>
-              </div>
-            </SectionBody>
-        </Section>  
-      {/if}
+      <UserDetails />
 
       {#if isWalletCorrect}
           <Section>
